A base docker image for my roon extension images
Dockerfile: Build the basic image complete with scripts to start the extension, and periodically (daily) check for updates. If updates are required node will be killed, and the updates will be installed when the container restarts. Use with the docker run option "--restart=unless-stopped", but remember to deal with garbage collection of exited images.

check-for-updates.sh:
update-and-start.sh: The scripts used by the dockerfile.