Timestamp all logs and fix the update checks
lib/check-for-updates.sh:
lib/update-and-start.sh: Move the scripts into a new libs directory and
call them from wrappers that ensure the output is timestamped. crond has
a background option, so use that rather than nohup.

bin/check-for-updates.sh:
bin/update-and-start.sh: Wrappers which run the implementation scripts
using tstamp. In the case of check-for-updates the output needs to be
redirected to a log file in the extensions volume so that it is
accessible after the container restarts; for update-and-start output to
stdout is fine as the docker logging will deal with that.

bin/tstamp: A wrapper around ts from moreutils to generate sensible
format timestamps.

Dockerfile: Updates to match. Note that moreutils is only in the
edge/testing Alpine repository.